tmp modify rms

+++ include/ops/bp/ew.cuh
@@ -127,7 +127,7 @@ public:
 
         const int tid = blockIdx.x * kBlockSize + threadIdx.x;
         for (int64_t i = tid; i < n; i += blockDim.x * gridDim.x) {
-          z[i] = pow(x[i], y[i]);
+          z[i] = powf(x[i], y[i]);
         }
     }
 };
